Move TestSuiteGeneratedByLLM into core module
