--- core/src/main/kotlin/org/jetbrains/research/testspark/core/parsing/test/ParsedTestSuite.kt
@@ -1,9 +1,0 @@
-package org.jetbrains.research.testspark.core.parsing.test
-
-data class ParsedTestSuite(
-    val imports: Set<String>,
-    val packageString: String,
-    val runWith: String,
-    val otherInfo: String,
-    val testCases: MutableList<TestCaseGeneratedByLLM>
-)
