+++ core/src/main/kotlin/org/jetbrains/research/testspark/core/test/TestSuiteGeneratedByLLM.kt
@@ -1,6 +1,4 @@
-package org.jetbrains.research.testspark.tools.llm.test
-
-import org.jetbrains.research.testspark.core.parsing.test.TestCaseGeneratedByLLM
+package org.jetbrains.research.testspark.core.test
 
 /**
  * Represents a test suite generated by LLM.
