+++ src/main/kotlin/org/jetbrains/research/testspark/tools/llm/generation/RequestManager.kt
@@ -7,7 +7,7 @@ import org.jetbrains.research.testspark.bundles.TestSparkBundle
 import org.jetbrains.research.testspark.tools.llm.SettingsArguments
 import org.jetbrains.research.testspark.tools.llm.error.LLMErrorManager
 import org.jetbrains.research.testspark.tools.llm.generation.openai.ChatMessage
-import org.jetbrains.research.testspark.tools.llm.test.TestSuiteGeneratedByLLM
+import org.jetbrains.research.testspark.core.test.TestSuiteGeneratedByLLM
 
 abstract class RequestManager {
     enum class SendResult { OK, TOOLONG, OTHER }
