+++ src/main/kotlin/org/jetbrains/research/testspark/tools/llm/test/TestSuitePresenter.kt
@@ -2,6 +2,7 @@ package org.jetbrains.research.testspark.tools.llm.test
 
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
+import org.jetbrains.research.testspark.core.test.TestSuiteGeneratedByLLM
 import org.jetbrains.research.testspark.services.JavaClassBuilderService
